+++ Stitch/Graph/LayerInspector/LayerInspectorPortView.swift
@@ -185,8 +185,9 @@ struct InspectorLayerInputView: View {
                                     document: document,
                                     inputField: inputFieldViewModel,
                                     node: node,
-                                    rowViewModel: inputRowViewModel,
-                                    canvasItem: nil,
+                                    rowId: inputRowViewModel.id,
+                                    layerInputPort: inputRowViewModel.layerInput,
+                                    canvasItemId: nil,
                                     rowObserver: inputRowObserver,
                                     isCanvasItemSelected: false,
                                     hasIncomingEdge: false,
@@ -241,8 +242,9 @@ struct LayerInputFieldsView: View {
                            document: document,
                            inputField: inputFieldViewModel,
                            node: node,
-                           rowViewModel: rowViewModel,
-                           canvasItem: canvasNode,
+                           rowId: rowViewModel.id,
+                           layerInputPort: rowViewModel.layerInput,
+                           canvasItemId: canvasNode.id,
                            rowObserver: rowObserver,
                            isCanvasItemSelected: isNodeSelected,
                            hasIncomingEdge: rowObserver.upstreamOutputCoordinate.isDefined,
